# Avoid cache miss when getting instruct func