… pkg/gb/instructions.go