This makes the instruction array a global variable. This improves performance by ~3ns/op. This also includes a new Benchmark for ExecuteNextOpcode()

Benchmarks:
```
OLD
goos: windows
goarch: amd64
pkg: github.com/Humpheh/goboy/pkg/gb
BenchmarkGameboy_ExecuteNextOpcode
2020/05/29 18:39:07 Cart type: 0x01 (MBC1)
2020/05/29 18:39:07 Cart type: 0x01 (MBC1)
2020/05/29 18:39:07 Cart type: 0x01 (MBC1)
2020/05/29 18:39:07 Cart type: 0x01 (MBC1)
2020/05/29 18:39:07 Cart type: 0x01 (MBC1)
BenchmarkGameboy_ExecuteNextOpcode-8    52322690                23.3 ns/op
PASS
ok      github.com/Humpheh/goboy/pkg/gb 1.418s
?       github.com/Humpheh/goboy/pkg/gb/io      [no test files]

NEW
goos: windows
goarch: amd64
pkg: github.com/Humpheh/goboy/pkg/gb
BenchmarkGameboy_ExecuteNextOpcode
2020/05/29 18:41:54 Cart type: 0x01 (MBC1)
2020/05/29 18:41:54 Cart type: 0x01 (MBC1)
2020/05/29 18:41:54 Cart type: 0x01 (MBC1)
2020/05/29 18:41:54 Cart type: 0x01 (MBC1)
2020/05/29 18:41:54 Cart type: 0x01 (MBC1)
BenchmarkGameboy_ExecuteNextOpcode-8    56845635                20.5 ns/op
PASS
ok      github.com/Humpheh/goboy/pkg/gb 1.362s
?       github.com/Humpheh/goboy/pkg/gb/io      [no test files]
```