@TotallyGamerJet 2020-06-09
It's all good. I'd kind of crazy right now. I also noticed that the secondary at of instructions are also stored in the gb. Perhaps you would also want to move those as well